gi/shadergen: fix emissive face indices not being 32 bit ints
Most scenes have more than 65535 triangles - although not all are
emissive, we store the INDEX, which can easily exceed 16 bit.

+++ src/shadergen/shaders/main.comp.hlsl
@@ -20,7 +20,7 @@ StructuredBuffer<face> faces;
 
 #ifdef NEXT_EVENT_ESTIMATION
 [[vk::binding(3)]]
-StructuredBuffer<uint16_t> emissive_face_indices;
+StructuredBuffer<uint32_t> emissive_face_indices;
 #endif
 
 [[vk::binding(4)]]
